+++ src/modules/tuner/tuner.ts
@@ -3,7 +3,7 @@ import { LiteEvent } from '../common/liteEvent';
 import { WebCoreConfig } from './config';
 import { Envir } from '../envir/envir';
 import { inject, injectable } from '../common/di';
-import { Environment } from '../envir/environment';
+import type { Environment } from '../envir/environment';
 
 const DEFAULT_CONFIG_NAME = 'default';