+++ vite.config.ts
@@ -1,6 +1,5 @@
 import { resolve } from 'path';
 import { defineConfig } from 'vitest/config';
-import typescript from '@rollup/plugin-typescript';
 import pkg from './package.json' with {type: 'json'};
 
 const tsConfig = {
@@ -36,11 +35,8 @@ export default defineConfig({
     },
     minify: false,
     sourcemap: true,
-    rollupOptions: {
+    rolldownOptions: {
       external: ['axios', 'qs'],
-      plugins: [
-        typescript(tsConfig),
-      ],
     },
   },
   test: {